Mahout 사용자 기반 추천 구현

+++ src/main/java/com/example/demo/model/UserPreference.java
@@ -26,6 +26,6 @@ public class UserPreference {
 
     //UserRecommendApiController의 dataToCSV의 로직에 의해 int를 float으로 바꿈
     @Column
-    private float preference;
+    private int preference;
 
 }
